✨ feat: 로그인 상태 확인 시 닉네임도 전송
