+++ packages/backend/src/auth/auth.controller.ts
@@ -1,16 +1,17 @@
-import { Controller, Post, Req, Res } from '@nestjs/common';
-import { ApiOperation, ApiTags } from '@nestjs/swagger';
+import { Controller, Get, Post, Req, Res } from '@nestjs/common';
+import { ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { Request, Response } from 'express';
 import { sessionConfig } from '@/configs/session.config';
+import { User } from '@/user/domain/user.entity';
 
 @ApiTags('Auth')
-@Controller('auth/logout')
-export class LogoutController {
+@Controller('auth')
+export class AuthController {
   @ApiOperation({
     summary: '로그아웃',
     description: '로그아웃을 진행한다.',
   })
-  @Post()
+  @Post('/logout')
   logout(@Req() req: Request, @Res() res: Response) {
     req.logout((err) => {
       if (err) {
@@ -28,5 +29,22 @@ export class LogoutController {
         return res.status(200).send({ message: 'Logged out successfully' });
       });
     });
+  }
+
+  @ApiOperation({
+    summary: '로그인 상태 확인',
+    description: '로그인 상태를 확인합니다.',
+  })
+  @ApiOkResponse({
+    description: '로그인된 상태',
+    example: { message: 'Authenticated' },
+  })
+  @Get('/status')
+  async user(@Req() request: Request) {
+    if (request.user) {
+      const user = request.user as User;
+      return { message: 'Authenticated', nickname: user.nickname };
+    }
+    return { message: 'Not Authenticated', nickname: null };
   }
 }
