+++ packages/backend/src/auth/auth.module.ts
@@ -3,7 +3,7 @@ import { PassportModule } from '@nestjs/passport';
 import { GoogleAuthController } from '@/auth/google/googleAuth.controller';
 import { GoogleAuthService } from '@/auth/google/googleAuth.service';
 import { GoogleStrategy } from '@/auth/google/strategy/google.strategy';
-import { LogoutController } from '@/auth/session/logout.controller';
+import { AuthController } from '@/auth/auth.controller';
 import { SessionSerializer } from '@/auth/session/session.serializer';
 import { TesterStrategy } from '@/auth/tester/strategy/tester.strategy';
 import { TesterAuthController } from '@/auth/tester/testerAuth.controller';
@@ -12,7 +12,7 @@ import { UserModule } from '@/user/user.module';
 
 @Module({
   imports: [UserModule, PassportModule.register({ session: true })],
-  controllers: [GoogleAuthController, TesterAuthController, LogoutController],
+  controllers: [GoogleAuthController, TesterAuthController, AuthController],
   providers: [
     GoogleStrategy,
     GoogleAuthService,
