+++ packages/backend/src/auth/google/googleAuth.controller.ts
@@ -1,6 +1,6 @@
-import { Controller, Get, Req, Res, UseGuards } from '@nestjs/common';
-import { ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
-import { Request, Response } from 'express';
+import { Controller, Get, Res, UseGuards } from '@nestjs/common';
+import { ApiOperation, ApiTags } from '@nestjs/swagger';
+import { Response } from 'express';
 import { GoogleAuthGuard } from '@/auth/google/guard/google.guard';
 
 @ApiTags('Auth')
@@ -22,21 +22,5 @@ export class GoogleAuthController {
   @UseGuards(GoogleAuthGuard)
   async handleRedirect(@Res() response: Response) {
     response.redirect('/');
-  }
-
-  @ApiOperation({
-    summary: '로그인 상태 확인',
-    description: '로그인 상태를 확인합니다.',
-  })
-  @ApiOkResponse({
-    description: '로그인된 상태',
-    example: { message: 'Authenticated' },
-  })
-  @Get('/status')
-  async user(@Req() request: Request) {
-    if (request.user) {
-      return { message: 'Authenticated' };
-    }
-    return { message: 'Not Authenticated' };
   }
 }
